tests: dbus_util module
